eyeflow-agent/requirements.txt
flask==2.3.3
python-socketio==5.9.0
python-engineio==4.7.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.3.0
psycopg2-binary==2.9.9
redis==5.0.0
APScheduler==3.10.4
pyyaml==6.0.1
